\name{infraFDTD.assist-package}
\alias{infraFDTD.assist-package}
\alias{infraFDTD.assist}
\docType{package}
\title{
\packageTitle{infraFDTD.assist}
}
\description{
\packageDescription{infraFDTD.assist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{infraFDTD.assist}
\packageIndices{infraFDTD.assist}

Before running the infraFDTD model, generate the input files using
PrepFiles. After running the model, binary snapshot files can be
converted into png files by MakeSnapshotInfo and MakeSnapshots.

}
\author{
\packageAuthor{infraFDTD.assist}

Maintainer: \packageMaintainer{infraFDTD.assist}
}
\references{
The infraFDTD code is available on request from Keehoon Kim at LLNL
(kim84@llnl.gov) and is described in Kim et al. (2015)
<doi:10.1002/2015GL064466>.
}
\keyword{ package }
\seealso{
  See package rNOMADS for help finding atmospheric profiles for given time/place.
}
